searchbar tests and component created
