+++ src/components/SearchBar.js
@@ -1,0 +1,12 @@
+
+function SearchBar() {
+
+    return (
+        <div className="searchBar">
+            <input type='search' placeholder='Search by Title'></input>
+            <button type='submit'>Search</button>
+        </div>
+    )
+}
+
+export default SearchBar;
